refactor unit test classes names to camel case move all unit test class files to progob package
